python-download-inta-file/requirements.txt
requests
peewee
PyMySQL
python-dateutil
beautifulsoup4
